+++ service/src/main/java/org/ehrbase/service/TemplateServiceImp.java
@@ -16,18 +16,29 @@
 
 package org.ehrbase.service;
 
+import com.nedap.archie.rm.composition.Composition;
 import org.apache.xmlbeans.XmlOptions;
 import org.ehrbase.api.definitions.OperationalTemplateFormat;
 import org.ehrbase.api.definitions.ServerConfig;
 import org.ehrbase.api.exception.InternalServerException;
 import org.ehrbase.api.exception.InvalidApiParameterException;
 import org.ehrbase.api.exception.ObjectNotFoundException;
+import org.ehrbase.api.service.CompositionService;
 import org.ehrbase.api.service.TemplateService;
+import org.ehrbase.building.webtemplateskeletnbuilder.WebTemplateSkeletonBuilder;
+import org.ehrbase.client.classgenerator.shareddefinition.Category;
+import org.ehrbase.client.classgenerator.shareddefinition.Language;
+import org.ehrbase.client.classgenerator.shareddefinition.Setting;
+import org.ehrbase.client.classgenerator.shareddefinition.Territory;
 import org.ehrbase.ehr.knowledge.TemplateMetaData;
+import org.ehrbase.openehr.sdk.examplegenerator.ExampleGeneratorConfig;
+import org.ehrbase.openehr.sdk.examplegenerator.ExampleGeneratorToCompositionWalker;
+import org.ehrbase.response.ehrscape.CompositionDto;
 import org.ehrbase.response.ehrscape.CompositionFormat;
 import org.ehrbase.response.ehrscape.StructuredString;
-import org.ehrbase.response.ehrscape.StructuredStringFormat;
 import org.ehrbase.response.ehrscape.TemplateMetaDataDto;
+import org.ehrbase.serialisation.walker.defaultvalues.DefaultValuePath;
+import org.ehrbase.serialisation.walker.defaultvalues.DefaultValues;
 import org.ehrbase.webtemplate.model.WebTemplate;
 import org.jooq.DSLContext;
 import org.openehr.schemas.v1.CARCHETYPEROOT;
@@ -38,6 +49,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.xml.namespace.QName;
 import java.nio.charset.StandardCharsets;
+import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -53,10 +65,12 @@ import java.util.stream.Collectors;
 public class TemplateServiceImp extends BaseServiceImp implements TemplateService {
 
     private final KnowledgeCacheService knowledgeCacheService;
+    private final CompositionService compositionService;
 
-    public TemplateServiceImp(KnowledgeCacheService knowledgeCacheService, DSLContext context, ServerConfig serverConfig) {
+    public TemplateServiceImp(KnowledgeCacheService knowledgeCacheService, DSLContext context, ServerConfig serverConfig, CompositionService compositionService) {
         super(knowledgeCacheService, context, serverConfig);
         this.knowledgeCacheService = Objects.requireNonNull(knowledgeCacheService);
+        this.compositionService = compositionService;
     }
 
     @Override
@@ -85,7 +99,24 @@ public class TemplateServiceImp extends BaseServiceImp implements TemplateServic
 
     @Override
     public StructuredString buildExample(String templateId, CompositionFormat format) {
-        return new StructuredString("", StructuredStringFormat.fromCompositionFormat(format));
+        WebTemplate webTemplate = findTemplate(templateId);
+        Composition composition = WebTemplateSkeletonBuilder.build(webTemplate, false);
+
+        ExampleGeneratorConfig object = new ExampleGeneratorConfig();
+
+        DefaultValues defaultValues = new DefaultValues();
+        defaultValues.addDefaultValue(DefaultValuePath.TIME, OffsetDateTime.now());
+        defaultValues.addDefaultValue(DefaultValuePath.LANGUAGE, Language.DE);
+        defaultValues.addDefaultValue(DefaultValuePath.TERRITORY, Territory.DE);
+        defaultValues.addDefaultValue(DefaultValuePath.SETTING, Setting.DENTAL_CARE);
+
+        composition.setCategory(Category.EVENT.toCodedText());
+
+        ExampleGeneratorToCompositionWalker walker = new ExampleGeneratorToCompositionWalker();
+        walker.walk(composition, object, webTemplate, defaultValues, templateId);
+
+        CompositionDto compositionDto = new CompositionDto(composition, templateId, null, null);
+        return compositionService.serialize(compositionDto, format);
     }
 
     @Override